+++ src/Core/Backend.Core.DDD/Events/IAsyncEventHandler.cs
@@ -1,9 +1,0 @@
-﻿using MediatR;
-
-namespace GoldenEye.Backend.Core.DDD.Events
-{
-    public interface IAsyncEventHandler<in TEvent> : IAsyncNotificationHandler<TEvent>
-           where TEvent : IEvent
-    {
-    }
-}