+++ src/Core/Backend.Core.DDD/Queries/IAsyncQueryHandler.cs
@@ -1,9 +1,0 @@
-﻿using MediatR;
-
-namespace GoldenEye.Backend.Core.DDD.Queries
-{
-    public interface IAsyncQueryHandler<in TQuery, TResponse> : IAsyncRequestHandler<TQuery, TResponse>
-           where TQuery : IQuery<TResponse>
-    {
-    }
-}