+++ src/Core/Backend.Core.DDD/Queries/IQueryHandler.cs
@@ -2,7 +2,7 @@
 
 namespace GoldenEye.Backend.Core.DDD.Queries
 {
-    public interface IQueryHandler<in TQuery, out TResponse> : IRequestHandler<TQuery, TResponse>
+    public interface IQueryHandler<in TQuery, TResponse> : IRequestHandler<TQuery, TResponse>
            where TQuery : IQuery<TResponse>
     {
     }