+++ src/Core/Backend.Core.DDD/Registration/Registration.cs
@@ -33,7 +33,7 @@ namespace GoldenEye.Backend.Core.DDD.Registration
 
         public static void AddEventStorePipeline(this IServiceCollection services)
         {
-            services.AddTransient(typeof(IAsyncNotificationHandler<>), typeof(EventStorePipeline<>));
+            services.AddTransient(typeof(INotificationHandler<>), typeof(EventStorePipeline<>));
         }
 
         public static void AddValidationPipeline(this IServiceCollection services)
@@ -50,15 +50,6 @@ namespace GoldenEye.Backend.Core.DDD.Registration
             services.AddTransient<ICommandHandler<TCommand>>(sp => sp.GetService<TCommandHandler>());
         }
 
-        public static void RegisterAsyncCommandHandler<TCommand, TCommandHandler>(this IServiceCollection services)
-            where TCommand : ICommand
-            where TCommandHandler : class, IAsyncCommandHandler<TCommand>
-        {
-            services.AddTransient<TCommandHandler>();
-            services.AddTransient<IAsyncRequestHandler<TCommand>>(sp => sp.GetService<TCommandHandler>());
-            services.AddTransient<IAsyncCommandHandler<TCommand>>(sp => sp.GetService<TCommandHandler>());
-        }
-
         public static void RegisterQueryHandler<TQuery, TResponse, TQueryHandler>(this IServiceCollection services)
             where TQuery : IQuery<TResponse>
             where TQueryHandler : class, IQueryHandler<TQuery, TResponse>
@@ -68,15 +59,6 @@ namespace GoldenEye.Backend.Core.DDD.Registration
             services.AddTransient<IQueryHandler<TQuery, TResponse>>(sp => sp.GetService<TQueryHandler>());
         }
 
-        public static void RegisterAsyncQueryHandler<TQuery, TResponse, TQueryHandler>(this IServiceCollection services)
-            where TQuery : IQuery<TResponse>
-            where TQueryHandler : class, IAsyncQueryHandler<TQuery, TResponse>
-        {
-            services.AddTransient<TQueryHandler>();
-            services.AddTransient<IAsyncRequestHandler<TQuery, TResponse>>(sp => sp.GetService<TQueryHandler>());
-            services.AddTransient<IAsyncQueryHandler<TQuery, TResponse>>(sp => sp.GetService<TQueryHandler>());
-        }
-
         public static void RegisterEventHandler<TEvent, TEventHandler>(this IServiceCollection services)
             where TEvent : IEvent
             where TEventHandler : class, IEventHandler<TEvent>
@@ -84,15 +66,6 @@ namespace GoldenEye.Backend.Core.DDD.Registration
             services.AddTransient<TEventHandler>();
             services.AddTransient<INotificationHandler<TEvent>>(sp => sp.GetService<TEventHandler>());
             services.AddTransient<IEventHandler<TEvent>>(sp => sp.GetService<TEventHandler>());
-        }
-
-        public static void RegisterAsyncEventHandler<TEvent, TEventHandler>(this IServiceCollection services)
-            where TEvent : IEvent
-            where TEventHandler : class, IAsyncEventHandler<TEvent>
-        {
-            services.AddTransient<TEventHandler>();
-            services.AddTransient<IAsyncNotificationHandler<TEvent>>(sp => sp.GetService<TEventHandler>());
-            services.AddTransient<IAsyncEventHandler<TEvent>>(sp => sp.GetService<TEventHandler>());
         }
     }
 }