+++ src/Sample/DDD/Backend.DDD.Sample/Changelog.md
@@ -1,1 +1,12 @@
-﻿
+﻿# v3.0.0 (29.12.2017) [Pull Request](https://github.com/oskardudycz/GoldenEye/pull/44)
+
+## Changes
+
+* updated packages version to most recent, breaking changes after migration to MediatR 4 **[MAJOR]**
+* added proper handling of `CancellationToken` for async methods in command and query handlers to be aligned with other async handling conventions **[MAJOR]**
+
+# v2.2.0 (18.12.2017)
+
+## Changes
+
+* added samples **[MAJOR]**